--- packages/effect/src/unstable/http/Multipasta.ts
@@ -1,14 +1,0 @@
-/**
- * Exposes the multipart parser used by the unstable HTTP modules.
- *
- * This module keeps the parser types and helpers available from the Effect HTTP
- * namespace without wrapping or changing them.
- *
- * @since 4.0.0
- */
-
-/**
- * @category multipart
- * @since 4.0.0
- */
-export * from "./Multipasta/internal/index.ts"
