--- packages/effect/src/unstable/http/Multipasta/HeadersParser.ts
@@ -1,14 +1,0 @@
-/**
- * Exposes the header parser used by multipart HTTP parsing.
- *
- * This module keeps the parser available from the Effect HTTP namespace without
- * wrapping or changing it.
- *
- * @since 4.0.0
- */
-
-/**
- * @category multipart
- * @since 4.0.0
- */
-export * from "./internal/headersParserApi.ts"
