--- packages/effect/src/unstable/http/Multipasta/Search.ts
@@ -1,14 +1,0 @@
-/**
- * Exposes the multipart search helpers.
- *
- * This module keeps the search helpers available from the Effect HTTP namespace
- * without wrapping or changing them.
- *
- * @since 4.0.0
- */
-
-/**
- * @category multipart
- * @since 4.0.0
- */
-export * from "./internal/searchApi.ts"
