--- packages/effect/src/unstable/http/Multipasta/Web.ts
@@ -1,14 +1,0 @@
-/**
- * Exposes the Web multipart helpers.
- *
- * This module keeps the Web-specific parser helpers available from the Effect
- * HTTP namespace without wrapping or changing them.
- *
- * @since 4.0.0
- */
-
-/**
- * @category multipart
- * @since 4.0.0
- */
-export * from "./internal/web.ts"
