+++ packages/effect/src/unstable/http/index.ts
@@ -127,7 +127,7 @@ export * as Multipart from "./Multipart.ts"
 /**
  * @since 4.0.0
  */
-export * as Multipasta from "./Multipasta.ts"
+export * as MultipartParser from "./MultipartParser.ts"
 
 /**
  * @since 4.0.0
